feat(login): 集成TapTap登录并重构用户ID管理
- 添加TapTap登录SDK依赖并实现异步登录流程
- 用UserLoginData结构统一存储用户登录信息
- 引入CRC64编码生成用户唯一ID
- 替换原有LoginServ.Uid为LoginServ.Id访问方式
- 修改相关模块（DBServ、GameSettingUI、PrefsServ、WorldServ）中用户ID引用
- 更新Android清单文件，添加INTERNET权限
- 清理无用测试代码空行
- 调整CRC64编码方法返回类型为long

+++ FLib/Sources/Encoder/CRC64.cs
@@ -35,7 +35,7 @@ namespace FLib
         /// <summary>
         /// 
         /// </summary>
-        public static ulong Encode(in string value)
+        public static long Encode(in string value)
         {
             var bytesCount = StringFLibUtility.Encoding.GetByteCount(value);
             if (bytesCount < 4096)
@@ -50,7 +50,7 @@ namespace FLib
         /// <summary>
         /// 
         /// </summary>
-        public static ulong Encode(ReadOnlySpan<byte> source, ulong crc = uint.MaxValue)
+        public static long Encode(ReadOnlySpan<byte> source, ulong crc = uint.MaxValue)
         {
             for (var i = 0; i < source.Length; i++)
             {
@@ -58,7 +58,7 @@ namespace FLib
                 idx ^= source[i];
                 crc = mLookup[idx] ^ (crc << 8);
             }
-            return crc;
+            return (long)crc;
         }
